Refined example
Intended to create zip safe example data to include in the binary
distribution. However, this defeated the purpose of creating files/file
structure that could be used in the example notebook. Additionally, it
violated many of the assumptions on building paths for the main script.
This commit undoes many of the file structure changes and explicitly
removes the attempts to make the example run out of the box.