Fix for the reverse/section matching
- The problem arose because the `map_from` section took on the
properties of the `map_to` section during section matching. This worked
in one relative orientation of the sections but not in the other. The
problem became apparent when trying to label the main tower of the LOVE
experiment, where the `map_from` segment had LAFs that were reversed
relative to the physical coordinates while the `map_to` segment
(unheated branch) had the physical coordinates aligned with the LAF
direction. This scenario is the opposite of the previous field campaign
data processing by some coincidence.
- The fix is in givng the `map_from` sections the same `reverse`
attribute as the `map_to` sections. In the section matching code the
`map_from` sections are flipped (e.g., LAFs decrease instead of
increase) and require no manual flipping when labeling as a result.